requirements.txt
pandas
tldextract
newspaper3k
ast
numpy
matplotlib
pywebcopy
urllib
re
requests
urlextract
retrying
unshortenit
time
openai
python-decouple